refactor: add report-card, comment-card, feed
